LLM_app/requirements.txt
fastapi
pydantic
langchain-groq
uvicorn[standard]